requirements.txt
# Core dependencies
faster-whisper>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
pynput>=1.7.6
PyGObject>=3.44.0
